+++ rediscluster/exceptions.py
@@ -85,3 +85,14 @@ class MasterDownError(ClusterDownError):
     """
     """
     pass
+
+
+class SlotNotCoveredError(RedisClusterException):
+    """
+    This error only happens in the case where the connection pool will try to
+    fetch what node that is covered by a given slot.
+
+    If this error is raised the client should drop the current node layout and
+    attempt to reconnect and refresh the node layout again
+    """
+    pass